requirements.txt
akshare==1.11.35
chinese_calendar==1.8.1
pandas==2.0.3
python_dateutil==2.8.2
pytz==2023.3.post1
pywencai==0.12.0
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
schedule==1.2.1
urllib3==1.26.16
